pyproject.toml
[project]
name = "logger-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Fabrizio Damicelli", email = "fabridamicelli@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "openai-whisper>=20250625",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
]

[build-system]
requires = ["uv_build>=0.11.2,<0.12.0"]
build-backend = "uv_build"
